bypassed youtube sign in issue

+++ server/services/audioService.js
@@ -3,7 +3,7 @@ import fs, { mkdirSync } from 'fs';
 import path, { dirname } from 'path';
 import ffmpegPath from 'ffmpeg-static';
 import { fileURLToPath } from 'url';
-
+import { fetchTranscript } from 'youtube-transcript';
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
@@ -13,38 +13,48 @@ if(!fs.existsSync(temp_dir)){
     fs.mkdirSync(temp_dir);
 }
 
-export const downloadAudio = async(videoId)=>{
+export const downloadAudio = async(videoId,transcription=true)=>{
     const videoUrl = `https://www.youtube.com/watch?v=${videoId}`;
-    const outputPath = path.join(temp_dir,`${videoId}.mp3`);
+    const textBlock = await fetchTranscript(videoUrl);
+    const fullText = textBlock.map((item)=> item.text).join(' ');
 
-    console.log("Starting to download audio for: ",videoId);
-    try {
-    await ytDlp(videoUrl, {
-      extractAudio: true,
-      audioFormat: 'mp3',
-      audioQuality: 9, 
-      limitRate: '2M',
-      postprocessorArgs: [
-        'ffmpeg:-threads 1 -b:a 32k'
-      ],
-      output: path.join(temp_dir, '%(id)s.%(ext)s'), 
-      noPlaylist: true,
-      ffmpegLocation: ffmpegPath,
-      jsRuntimes: 'node', 
-      cookies: path.join(__dirname,'../cookies.txt'),
-      extractorArgs: 'youtube:player_client=default,web_embedded'
-    })
-        
-    console.log("Audio download complete.For: ",outputPath);
-    return outputPath;
+    if(!transcription)
+    {
+        const outputPath = path.join(temp_dir,`${videoId}.mp3`);
+        console.log("Starting to download audio for: ",videoId);
+        try {
+            await ytDlp(videoUrl, {
+            extractAudio: true,
+            audioFormat: 'mp3',
+            audioQuality: 9, 
+            limitRate: '2M',
+            postprocessorArgs: [
+                'ffmpeg:-threads 1 -b:a 32k'
+            ],
+            output: path.join(temp_dir, '%(id)s.%(ext)s'), 
+            noPlaylist: true,
+            ffmpegLocation: ffmpegPath,
+            jsRuntimes: 'node', 
+            cookies: path.join(__dirname,'../cookies.txt'),
+            extractorArgs: 'youtube:player_client=default,web_embedded'
+            })
+                
+            console.log("Audio download complete.For: ",outputPath);
+            return outputPath;
 
-    } catch (error) {
-        console.log("Audio download failed: ",error.message);
-        throw error;
+            } catch (error) {
+                console.log("Audio download failed: ",error.message);
+                throw error;
+            }
     }
+
+     if(transcription){
+                return fullText;
+        }
 }
 
-export const deleteAudioFile = async(filepath)=>{
+export const deleteAudioFile = async(filepath,transcription=true)=>{
+   if(!transcription){
     try {
         if(fs.existsSync(filepath)){
             fs.unlinkSync(filepath);
@@ -53,5 +63,6 @@ export const deleteAudioFile = async(filepath)=>{
     } catch (error) {
         console.log("Error in deleting file: ",error.message);
     }
+   }
 
 }
